requirements.txt
flask==3.0.1
requests==2.32.3
jsonify==0.5
python-dotenv==1.1.0
sqlalchemy==2.0.41
pymysql==1.1.1
cryptography==41.0.7
